fix(smart-wallet): move types into a separate file
