+++ packages/client/wallets/smart-wallet/src/evm/transfer.ts
@@ -1,55 +1,8 @@
-import { type Abi, type Address, erc20Abi, erc721Abi } from "viem";
+import { type Abi, erc20Abi, erc721Abi } from "viem";
 
 import erc1155Abi from "@/abi/erc1155";
 
-import type { SmartWalletChain } from "./chains";
-
-interface EVMToken {
-    chain: SmartWalletChain;
-    contractAddress: Address;
-}
-
-interface NFTEVMToken extends EVMToken {
-    tokenId: string;
-    type: "nft";
-}
-
-interface SFTEVMToken extends EVMToken {
-    tokenId: string;
-    type: "sft";
-}
-
-interface ERC2OEVMToken extends EVMToken {
-    type: "ft";
-}
-
-export interface ERC20TransferType {
-    token: ERC2OEVMToken;
-    amount: bigint;
-}
-export interface SFTTransferType {
-    token: SFTEVMToken;
-    quantity: number;
-}
-export interface NFTTransferType {
-    token: NFTEVMToken;
-}
-export type TransferType = ERC20TransferType | SFTTransferType | NFTTransferType;
-
-interface TransferInputParams {
-    from: Address;
-    contract: Address;
-    to: Address;
-    config: TransferType;
-}
-
-interface TransferSimulationParams {
-    address: Address;
-    abi: Abi;
-    functionName: string;
-    args: unknown[];
-    tokenId?: string;
-}
+import type { ERC20TransferType, SFTTransferType, TransferInputParams, TransferSimulationParams } from "@/types/transfer";
 
 export function transferParams({ contract, config, from, to }: TransferInputParams): TransferSimulationParams {
     switch (config.token.type) {
